refactor: move utilities around
