--- pkg/semantic/compare.go
@@ -1,27 +1,0 @@
-package semantic
-
-import (
-	"math/big"
-)
-
-func convertToBigInt(str string) (*big.Int, bool) {
-	i, ok := new(big.Int).SetString(str, 10)
-
-	return i, ok
-}
-
-func minInt(x, y int) int {
-	if x > y {
-		return y
-	}
-
-	return x
-}
-
-func maxInt(x, y int) int {
-	if x < y {
-		return y
-	}
-
-	return x
-}
